+++ is_matrix_forge/led_matrix/display/helpers/columns.py
@@ -15,8 +15,10 @@ Description:
     $DESCRIPTION
 
 """
-from is_matrix_forge.led_matrix.commands.map import CommandVals
-from is_matrix_forge.led_matrix.constants import FWK_MAGIC
+from is_matrix_forge.led_matrix.hardware import (
+    commit_framebuffer_brightness,
+    stage_framebuffer_brightness_column,
+)
 from is_matrix_forge.log_engine import ROOT_LOGGER
 
 
@@ -26,13 +28,11 @@ MOD_LOGGER = ROOT_LOGGER.get_child('led_matrix.display.helpers.columns')
 def send_col(dev, s, x, vals):
     """Stage greyscale values for a single column. Must be committed with commit_cols()."""
     log = MOD_LOGGER.get_child('send_col')
-    command = FWK_MAGIC + [CommandVals.StageGreyCol, x] + vals
-    log.debug(f'Sending command: {command}')
-    s.write(bytes(command))
+    log.debug(f'Staging grayscale brightness column {x}')
+    stage_framebuffer_brightness_column(s, x, vals)
 
 
 def commit_cols(dev, s):
     """Commit the changes from sending individual cols with send_col(), displaying the matrix.
     This makes sure that the matrix isn't partially updated."""
-    command = FWK_MAGIC + [CommandVals.DrawGreyColBuffer, 0x00]
-    s.write(bytes(command))
+    commit_framebuffer_brightness(s)