+++ is_matrix_forge/led_matrix/errors/__init__.py
@@ -1,6 +1,7 @@
 from .base import LEDMatrixControllerError
 from .grid import GridDefinitionError, MalformedGridError
 from .matrix import (
+    FramebufferStateUnknownError,
     InvalidBrightnessError,
     MatrixConnectionError,
     MatrixError,
@@ -10,6 +11,7 @@ from .misc import ImplicitNameDerivationError, MiscError
 
 __all__ = [
     'GridDefinitionError',
+    'FramebufferStateUnknownError',
     'ImplicitNameDerivationError',
     'InvalidBrightnessError',
     'LEDMatrixControllerError',